% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{keybd.release}
\alias{keybd.release}
\title{Simulate Key Release}
\usage{
keybd.release(button)
}
\arguments{
\item{button}{character. The key release to simulate (not case sensitive).}
}
\description{
Simulates the release of keyboard keys held by \code{\link[KeyboardSimulator]{keybd.press}}. Multiple keys can be released simultaneously by using a \code{+} separator (see Examples). See \code{\link[KeyboardSimulator]{keyboard_value}} for supported keys.
}
\examples{
\dontrun{

# Move to the third working window
keybd.press('Alt', hold = TRUE)
keybd.press('Tab')
Sys.sleep(0.1)
keybd.press('Tab')
keybd.release('Alt')
}
}
\seealso{
\code{\link[KeyboardSimulator]{keybd.press}}
}
