% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{keybd.type_string}
\alias{keybd.type_string}
\title{Type a raw string}
\usage{
keybd.type_string(string)
}
\arguments{
\item{string}{character. The string expected to output (case sensitive).}
}
\description{
Type a raw string base on a given string.
}
\examples{
\dontrun{

# Type 'Hello world!'
keybd.type_string("Hello world!")
}
}
