% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{KeyboardSimulator-package}
\alias{KeyboardSimulator}
\alias{KeyboardSimulator-package}
\title{Keyboard and Mouse Input Simulation for Windows OS}
\description{
Control your keyboard and mouse with R code by simulating key presses and mouse clicks. The input simulation is implemented with the Windows API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChiHangChen/KeyboardSimulator}
  \item Report bugs at \url{https://github.com/ChiHangChen/KeyboardSimulator/issues}
}

}
\author{
\strong{Maintainer}: Jim Chen \email{jim71183@gmail.com}

Authors:
\itemize{
  \item Jeff Keller \email{github@vtkellers.com} [contributor]
}

Other contributors:
\itemize{
  \item Garry Hopwood \email{garryhopwood@hotmail.com} [contributor]
  \item Chieh Hsu \email{tiff103012345@gmail.com} [contributor]
}

}
