% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{KeyboardSimulator}
\alias{KeyboardSimulator}
\title{Keyboard and Mouse Input Simulation for Windows OS}
\description{
Control your keyboard and mouse with R code by simulating key presses and mouse clicks. The input simulation is implemented with the Windows API.
}
\author{
Jim Chen, Jeff Keller
}
