% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.click}
\alias{mouse.click}
\title{Simulate mouse click}
\usage{
mouse.click(button = "left", hold = FALSE)
}
\arguments{
\item{button}{character. The allowed values are "\code{left}" and "\code{right}"}

\item{hold}{logical. Should the key to be held after click.}
}
\description{
Simulate mouse click
}
\examples{
mouse.click(button = "left", hold = TRUE)
}
