% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.get_cursor}
\alias{mouse.get_cursor}
\title{Get Current Cursor Coordinate}
\usage{
mouse.get_cursor()
}
\description{
Get current cursor coordinate of screen .
}
\examples{
\dontrun{

mouse.get_cursor()
}
}
