% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.move}
\alias{mouse.move}
\title{Move Cursor to Specific Location}
\usage{
mouse.move(x, y)
}
\arguments{
\item{x}{numeric. X-axis of screen.}

\item{y}{numeric. Y-axis of screen.}
}
\description{
Move cursor to specific coordinate of screen .
}
\examples{
\dontrun{

# Move cursor to middle of screen in 1080FHD monitor
mouse.move(x=960,y=540)
}
}
