% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.release}
\alias{mouse.release}
\title{Release Held key clicked by \code{\link[KeyboardSimulator]{mouse.click}}}
\usage{
mouse.release(button = "left")
}
\arguments{
\item{button}{character. The allowed values are "\code{left}" and "\code{right}"}
}
\description{
Release Held key clicked by \code{\link[KeyboardSimulator]{mouse.click}}
}
\examples{
\dontrun{

## Click right button
mouse.click(button = "right", hold = TRUE)
mouse.release(button = "right")
}
}
