% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_value.R
\docType{data}
\name{keyboard_value}
\alias{keyboard_value}
\title{keyboard_value}
\format{
An object of class \code{data.frame} with 79 rows and 4 columns.
}
\usage{
keyboard_value
}
\description{
List of supported keyboard keys, along with their virtual-key and hardware scan codes. A field indicating whether the key is an "extended key" with a \code{0xE0} prefix byte is also included to differentiate between duplicate keys on the keyboard and num pad. For example, while the \code{1} key usually behaves the same as the \code{1} key on the num pad, some applications see these as two distinct keys.
}
\details{
Supported keys:

\itemize{
 \item{a}
 \item{b}
 \item{c}
 \item{d}
 \item{e}
 \item{f}
 \item{g}
 \item{h}
 \item{i}
 \item{j}
 \item{k}
 \item{l}
 \item{m}
 \item{n}
 \item{o}
 \item{p}
 \item{q}
 \item{r}
 \item{s}
 \item{t}
 \item{u}
 \item{v}
 \item{w}
 \item{x}
 \item{y}
 \item{z}
 \item{0}
 \item{1}
 \item{2}
 \item{3}
 \item{4}
 \item{5}
 \item{6}
 \item{7}
 \item{8}
 \item{9}
 \item{num0}
 \item{num1}
 \item{num2}
 \item{num3}
 \item{num4}
 \item{num5}
 \item{num6}
 \item{num7}
 \item{num8}
 \item{num9}
 \item{f1}
 \item{f2}
 \item{f3}
 \item{f4}
 \item{f5}
 \item{f6}
 \item{f7}
 \item{f8}
 \item{f9}
 \item{f10}
 \item{f11}
 \item{f12}
 \item{backspace}
 \item{tab}
 \item{enter}
 \item{shift}
 \item{ctrl}
 \item{alt}
 \item{capslock}
 \item{esc}
 \item{pageup}
 \item{pagedown}
 \item{end}
 \item{home}
 \item{left}
 \item{up}
 \item{right}
 \item{down}
 \item{insert}
 \item{space}
 \item{del}
 \item{numlock}
 \item{win}
}
}
\references{
\href{http://www.philipstorr.id.au/pcbook/book3/scancode.htm}{Hardware Scan Codes}, \href{https://docs.microsoft.com/en-us/windows/desktop/inputdev/virtual-key-codes}{Virtual-Key Codes}
}
\keyword{data}
