% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.click}
\alias{mouse.click}
\title{Simulate mouse click}
\usage{
mouse.click(button = "left", hold = F)
}
\arguments{
\item{button}{character. The key to be released, can release several keys simultaneously with '+' as separator
, uppercase or lowercase are not sensitive.}
}
\description{
Simulate mouse click
}
\examples{
## Move to the third working window
}
