% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.click}
\alias{mouse.click}
\title{Simulate Mouse Clicks}
\usage{
mouse.click(button = "left", hold = FALSE)
}
\arguments{
\item{button}{character. Allowed values are \code{"left"}, \code{"right"}, and \code{"middle"}.}

\item{hold}{logical. Whether the button should be held down.}
}
\description{
Simulate left, right, and middle mouse clicks.
}
\examples{
\dontrun{

# left mouse click
mouse.click(button = "left")

# left mouse click and hold
mouse.click(button = "left", hold = TRUE)
}
}
\seealso{
\code{\link[KeyboardSimulator]{mouse.release}}
}
