% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.move}
\alias{mouse.move}
\title{Move Cursor to Specific Location}
\usage{
mouse.move(x, y, duration = NA, step_ratio = 0.01)
}
\arguments{
\item{x}{numeric. X-axis of screen.}

\item{y}{numeric. Y-axis of screen.}

\item{duration}{numeric. Cursor movement time in seconds, there might be some delay on different computer.}

\item{step_ratio}{numeric. Ratio of total distance in each step, only available when \code{duration} is not \code{NA}.}
}
\description{
Move cursor to specific coordinate of screen .
}
\examples{
\dontrun{

# Move cursor to middle of screen in 1080FHD monitor
mouse.move(x = 960, y = 540)

# Move cursor to middle of screen in 1080FHD monitor within 3 seconds
mouse.move(x = 960, y = 540, duration = 3)
}
}
