% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mouse.release}
\alias{mouse.release}
\title{Simulate Mouse Click Release}
\usage{
mouse.release(button = "left")
}
\arguments{
\item{button}{character. Allowed values are "\code{left}" and "\code{right}".}
}
\description{
Simulate the release of mouse button held by \code{\link[KeyboardSimulator]{mouse.click}}.
}
\examples{
\dontrun{

# right mouse click and hold
mouse.click(button = "right", hold = TRUE)

# release right click
mouse.release(button = "right")
}
}
\seealso{
\code{\link[KeyboardSimulator]{mouse.click}}
}
